+++ src/main/java/study/posthub/global/security/oauth2/CustomOAuth2Member.java
@@ -6,6 +6,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import study.posthub.domain.member.entity.Authority;
+import study.posthub.global.security.oauth2.dto.OAuth2Response;
 
 import java.util.Collection;
 import java.util.Collections;
