+++ src/main/java/study/posthub/global/security/oauth2/dto/GoogleResponse.java
@@ -1,4 +1,4 @@
-package study.posthub.global.security.oauth2;
+package study.posthub.global.security.oauth2.dto;
 
 import lombok.RequiredArgsConstructor;
