+++ src/main/java/study/posthub/global/security/oauth2/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package study.posthub.global.security.oauth2;
+package study.posthub.global.security.oauth2.dto;
 
 public interface OAuth2Response {
